Fix path.exists output
as of penlight 1.14.0, `path.exists(P)` return `P` if it exists, `false` otherwise. In particular, `path.exists(P)==true` is always false.

The documentation of penlight 1.14.0 states that the return value is either `P` or `nil` which is buggy. When this is fixed, the definition will need to be adapted.